examples/chained-apps-with-otel/app3/requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
